requirements.txt
asgiref==3.8.1
Django==5.1.2
django-currentuser==0.8.0
django-modeltranslation==0.19.10
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
